fix: resolve compilation warnings and extract templates to external files
- Fix serialVersionUID warnings by adding serialVersionUID to all serializable classes
- Fix unchecked cast warnings with proper @SuppressWarnings annotations
- Fix lossy cast warnings with explicit (int) casts
- Resolve trailing whitespace issues by extracting HTML templates to external .ftl files
- Add proper exception handling for TemplateProcessingException and ExportException
- Update test files to use correct API methods and enum values
- Replace problematic text blocks with external template loading for better maintainability

All main source code now compiles cleanly with zero warnings under -Werror flag.

+++ wasmtime4j-comparison-tests/src/main/java/ai/tegmentum/wasmtime4j/comparison/reporters/ComparisonSummary.java
@@ -256,15 +256,15 @@ public final class ComparisonSummary {
       // Count issues by severity (simplified categorization)
       final long criticalDiscrepancies = result.getCriticalDiscrepancyCount();
       if (criticalDiscrepancies > 0) {
-        highPriorityIssues += criticalDiscrepancies;
+        highPriorityIssues += (int) criticalDiscrepancies;
       }
 
       // Additional discrepancies classified as medium/low priority
       final int totalDiscrepancies = result.getDiscrepancies().size();
       final long remainingDiscrepancies = totalDiscrepancies - criticalDiscrepancies;
       if (remainingDiscrepancies > 0) {
-        mediumPriorityIssues += remainingDiscrepancies / 2;
-        lowPriorityIssues += remainingDiscrepancies - (remainingDiscrepancies / 2);
+        mediumPriorityIssues += (int) (remainingDiscrepancies / 2);
+        lowPriorityIssues += (int) (remainingDiscrepancies - (remainingDiscrepancies / 2));
       }
     }
